+++ aristonremotethermo/ariston.py
@@ -111,7 +111,7 @@ class AristonHandler:
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     """
 
-    _VERSION = "1.0.35"
+    _VERSION = "1.0.37"
 
     _LOGGER = logging.getLogger(__name__)
     _LEVEL_CRITICAL = "CRITICAL"
@@ -209,6 +209,7 @@ class AristonHandler:
     _VAL_SUMMER = "summer"
     _VAL_OFF = "off"
     _VAL_HEATING_ONLY = "heating_only"
+    _VAL_COOLING = "cooling"
     _VAL_MANUAL = "manual"
     _VAL_PROGRAM = "program"
     _VAL_UNKNOWN = "unknown"
@@ -248,7 +249,7 @@ class AristonHandler:
     _HTTP_PARAM_DELAY = 30.0
 
     # Conversions between parameters
-    _MODE_TO_VALUE = {_VAL_WINTER: 1, _VAL_SUMMER: 0, _VAL_OFF: 5, _VAL_HEATING_ONLY: 2}
+    _MODE_TO_VALUE = {_VAL_WINTER: 1, _VAL_SUMMER: 0, _VAL_OFF: 5, _VAL_HEATING_ONLY: 2, _VAL_COOLING: 3}
     _VALUE_TO_MODE = {value: key for (key, value) in _MODE_TO_VALUE.items()}
 
     _CH_MODE_TO_VALUE = {_VAL_MANUAL: 2, _VAL_PROGRAM: 3, _VAL_UNKNOWN: 0}
@@ -1342,19 +1343,27 @@ class AristonHandler:
                     self._ariston_sensors[self._PARAM_CH_FLAME][self._VALUE] = None
 
                 try:
-                    if not self._ariston_data["zone"]["heatRequest"] and self._ariston_data["flameSensor"]:
-                        self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = True
-                    elif self._ariston_data["flameForDhw"]:
-                        self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = True
-                    elif self._ariston_data["dhwStorageTemp"] != self._INVALID_STORAGE_TEMP \
-                            and self._dhw_trend_up \
-                            and self._VALUE_TO_MODE[self._ariston_data["mode"]] in [self._VAL_SUMMER, self._VAL_WINTER]\
-                            and self._ariston_data["flameSensor"]:
-                        self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = True
-                    elif self._dhw_unknown_as_on and self._ariston_data["flameSensor"]:
-                        self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = True
+                    if self._ariston_data["dhwStorageTemp"] != self._INVALID_STORAGE_TEMP:
+                        # trend can be used
+                        _dhw_automation_possible = True
                     else:
-                        self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = False
+                        _dhw_automation_possible = False
+                    _dhw_flame = False
+                    if self._ariston_data["flameSensor"] and \
+                        self._VALUE_TO_MODE[self._ariston_data["mode"]] in [self._VAL_SUMMER, self._VAL_WINTER]:
+                        if not self._ariston_data["zone"]["heatRequest"]:
+                            # flame is not CH meaning it is DHW
+                            _dhw_flame = True
+                        elif self._ariston_data["flameForDhw"]:
+                            # wow, does DHW flag actually work for someone
+                            _dhw_flame = True
+                        elif _dhw_automation_possible and self._dhw_trend_up:
+                            # temeparture increases, assume flame on
+                            _dhw_flame = True
+                        elif not _dhw_automation_possible and self._dhw_unknown_as_on:
+                            # no automation, use default value
+                            _dhw_flame = True
+                    self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = _dhw_flame
                 except KeyError:
                     self._ariston_sensors[self._PARAM_DHW_FLAME][self._VALUE] = None